+++ src/lib/sdkWebWalletCrossWindowProvider.ts
@@ -1,2 +1,1 @@
 export { CrossWindowProvider } from '@multiversx/sdk-web-wallet-cross-window-provider/out/CrossWindowProvider/CrossWindowProvider';
-export { MetamaskProxyProvider } from '@multiversx/sdk-metamask-proxy-provider/out/MetamaskProxyProvider/MetamaskProxyProvider';