+++ src/passkeyProvider/lib/webauthn-prf/index.ts
@@ -1,5 +1,0 @@
-import * as client from "./client.js";
-import * as utils from "./utils.js";
-
-export { client, utils };
-export default { client, utils };