+++ src/passkeyProvider/operation.ts
@@ -1,7 +1,0 @@
-export enum Operation {
-    Connect = "connect",
-    Logout = "logout",
-    SignTransactions = "signTransactions",
-    SignMessage = "signMessage",
-    CancelAction = "cancelAction",
-}