+++ src/providers/accountProvider.ts
@@ -2,12 +2,10 @@ import { ExtensionProvider } from '@multiversx/sdk-extension-provider';
 import { HWProvider } from '@multiversx/sdk-hw-provider';
 import { MetamaskProvider } from '@multiversx/sdk-metamask-provider/out/metamaskProvider';
 import { OperaProvider } from '@multiversx/sdk-opera-provider';
+import { PasskeyProvider } from '@multiversx/sdk-passkey-provider/out';
+import { IframeProvider } from '@multiversx/sdk-web-wallet-iframe-provider/out';
 import { WalletProvider } from '@multiversx/sdk-web-wallet-provider';
-import {
-  CrossWindowProvider,
-  MetamaskProxyProvider
-} from 'lib/sdkWebWalletCrossWindowProvider';
-import { PasskeyProvider } from 'passkeyProvider';
+import { CrossWindowProvider } from 'lib/sdkWebWalletCrossWindowProvider';
 import { IDappProvider } from 'types';
 import { WalletConnectV2Provider } from 'utils/walletconnect/__sdkWalletconnectProvider';
 import { emptyProvider } from './utils';
@@ -19,7 +17,7 @@ export type ProvidersType =
   | MetamaskProvider
   | WalletProvider
   | CrossWindowProvider
-  | MetamaskProxyProvider
+  | IframeProvider
   | HWProvider
   | OperaProvider
   | WalletConnectV2Provider;